opencdx-protector/src/main/java/cdx/opencdx/protector/service/impl/ProtectorServiceImpl.java
/*
 * Copyright 2023 Safe Health Systems, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package cdx.opencdx.protector.service.impl;

import cdx.opencdx.commons.service.OpenCDXAuditService;
import cdx.opencdx.commons.service.OpenCDXCurrentUser;
import cdx.opencdx.grpc.neural.protector.AnomalyDetectionDataRequest;
import cdx.opencdx.grpc.neural.protector.AuthorizationControlDataRequest;
import cdx.opencdx.grpc.neural.protector.PrivacyProtectionDataRequest;
import cdx.opencdx.grpc.neural.protector.RealTimeMonitoringDataRequest;
import cdx.opencdx.grpc.neural.protector.SecurityResponse;
import cdx.opencdx.grpc.neural.protector.UserBehaviorAnalysisDataRequest;
import cdx.opencdx.protector.service.ProtectorService;
import io.micrometer.observation.annotation.Observed;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * Service for processing protector Requests
 */
@Service
@Observed(name = "opencdx")
public class ProtectorServiceImpl implements ProtectorService {

    /**
     * Constructor taking the a PersonRepository
     *
     * @param personRepository    repository for interacting with the database.
     * @param openCDXAuditService Audit service for tracking FDA requirements
     * @param openCDXCurrentUser Current User Service.
     */
    @Autowired
    public ProtectorServiceImpl(OpenCDXAuditService openCDXAuditService, OpenCDXCurrentUser openCDXCurrentUser) {}

    /**
     * Process the AnomalyDetectionData
     * @param request request the process
     * @return Message generated for this request.
     */
    @Override
    public SecurityResponse detectAnomalies(AnomalyDetectionDataRequest request) {
        // placeholder
        return SecurityResponse.newBuilder()
                .setResponse("SecurityResponse [detectAnomalies]")
                .build();
    }

    /**
     * Process the AuthorizationControlData
     * @param request request the process
     * @return Message generated for this request.
     */
    @Override
    public SecurityResponse enforceAuthorizationControl(AuthorizationControlDataRequest request) {
        return SecurityResponse.newBuilder()
                .setResponse("SecurityResponse [enforceAuthorizationControl]")
                .build();
    }

    /**
     * Process the PrivacyProtectionData
     * @param request request the process
     * @return Message generated for this request.
     */
    @Override
    public SecurityResponse protectPrivacy(PrivacyProtectionDataRequest request) {
        // placeholder
        return SecurityResponse.newBuilder()
                .setResponse("SecurityResponse [protectPrivacy]")
                .build();
    }

    /**
     * Process the RealTimeMonitoringData
     * @param request request the process
     * @return Message generated for this request.
     */
    @Override
    public SecurityResponse monitorRealTimeActivity(RealTimeMonitoringDataRequest request) {
        // placeholder
        return SecurityResponse.newBuilder()
                .setResponse("SecurityResponse [monitorRealTimeActivity]")
                .build();
    }

    /**
     * Process the UserBehaviorAnalysisData
     * @param request request the process
     * @return Message generated for this request.
     */
    @Override
    public SecurityResponse analyzeUserBehavior(UserBehaviorAnalysisDataRequest request) {
        // placeholder
        return SecurityResponse.newBuilder()
                .setResponse("SecurityResponse [analyzeUserBehavior]")
                .build();
    }
}
